new month schema

+++ src/main/java/com/payroll/backend/assignment/AssignmentRequest.java
@@ -2,12 +2,13 @@ package com.payroll.backend.assignment;
 
 import lombok.Getter;
 import lombok.Setter;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
 public class AssignmentRequest {
     private Long userId;
-    private Double percentage;
+    private BigDecimal percentage;
     private PayoutBasis basisType;
     private Long sourceUserId;
 }
